requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.1.0
langchain-huggingface==0.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3